Add support for Authorize.net Accept.js
