+++ jazkarta/shop/profiles/plone5/metadata.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <metadata>
-  <version>10</version>
+  <version>11</version>
   <dependencies>
   	<dependency>profile-collective.z3cform.datagridfield:default</dependency>
   </dependencies>
